+++ crystalfontz/keys.py
@@ -1,6 +1,6 @@
 from dataclasses import asdict, dataclass
 from enum import Enum
-from typing import List, Literal, Self, Type
+from typing import List, Literal, Self, Tuple, Type
 
 KeyPress = (
     Literal[0x01]
@@ -19,6 +19,17 @@ KP_RIGHT: KeyPress = 0x10
 KP_DOWN: KeyPress = 0x20
 
 
+def keypress_repr(keypress: KeyPress) -> str:
+    return {
+        KP_UP: "KP_UP",
+        KP_ENTER: "KP_ENTER",
+        KP_EXIT: "KP_EXIT",
+        KP_LEFT: "KP_LEFT",
+        KP_RIGHT: "KP_RIGHT",
+        KP_DOWN: "KP_DOWN",
+    }[keypress]
+
+
 @dataclass
 class KeyState:
     """
@@ -31,6 +42,7 @@ class KeyState:
                                poll.
     """
 
+    keypress: KeyPress
     pressed: bool
     pressed_since: bool
     released_since: bool
@@ -42,12 +54,18 @@ class KeyState:
         released_since = state[2]
 
         return cls(
+            keypress=keypress,
             pressed=bool(pressed & keypress),
             pressed_since=bool(pressed_since & keypress),
             released_since=bool(released_since & keypress),
         )
 
-    def to_bytes(self: Self, keypress: KeyPress) -> bytes: ...
+    def to_bytes(self: Self) -> Tuple[int, int, int]:
+        pressed = self.keypress if self.pressed else 0x00
+        pressed_since = self.keypress if self.pressed_since else 0x00
+        released_since = self.keypress if self.released_since else 0x00
+
+        return (pressed, pressed_since, released_since)
 
 
 @dataclass
@@ -87,28 +105,34 @@ class KeyStates:
         pressed_since = 0x00
         released_since = 0x00
 
-        for state, keypress in [
-            (self.up, KP_UP),
-            (self.enter, KP_ENTER),
-            (self.exit, KP_EXIT),
-            (self.left, KP_LEFT),
-            (self.right, KP_RIGHT),
-            (self.down, KP_DOWN),
+        for state in [
+            self.up,
+            self.enter,
+            self.exit,
+            self.left,
+            self.right,
+            self.down,
         ]:
-            pressed = (pressed ^ keypress) if state.pressed else pressed
-            pressed_since = (
-                (pressed_since ^ keypress) if state.pressed_since else pressed_since
-            )
-            released_since = (
-                (released_since ^ keypress) if state.released_since else released_since
-            )
+            p, p_s, r_s = state.to_bytes()
+            pressed = pressed ^ p
+            pressed_since = pressed_since ^ p_s
+            released_since = released_since ^ r_s
 
         return bytes([pressed, pressed_since, released_since])
 
     def __repr__(self: Self) -> str:
         repr_ = ""
         for name, state in asdict(self).items():
-            st = ", ".join([f"{n}={'yes' if s else 'no'}" for n, s in state.items()])
+            st = ", ".join(
+                [
+                    (
+                        f"{n}={'yes' if s else 'no'}"
+                        if isinstance(s, bool)
+                        else f"{n}={keypress_repr(s)}"
+                    )
+                    for n, s in state.items()
+                ]
+            )
             repr_ += f"{name}: {st}\n"
         return repr_[0:-1]