+++ tests/test_device.py
@@ -2,15 +2,36 @@ import pytest
 
 from crystalfontz.atx import AtxPowerSwitchFunction, AtxPowerSwitchFunctionalitySettings
 from crystalfontz.device import CFA533, CFA533Status, Device, DeviceStatus
-from crystalfontz.keys import KeyState, KeyStates
+from crystalfontz.keys import (
+    KeyState,
+    KeyStates,
+    KP_DOWN,
+    KP_ENTER,
+    KP_EXIT,
+    KP_LEFT,
+    KP_RIGHT,
+    KP_UP,
+)
 
 KEY_STATES = KeyStates(
-    up=KeyState(pressed=False, pressed_since=False, released_since=False),
-    enter=KeyState(pressed=True, pressed_since=False, released_since=False),
-    exit=KeyState(pressed=False, pressed_since=True, released_since=False),
-    left=KeyState(pressed=False, pressed_since=False, released_since=True),
-    right=KeyState(pressed=True, pressed_since=True, released_since=False),
-    down=KeyState(pressed=True, pressed_since=False, released_since=True),
+    up=KeyState(
+        keypress=KP_UP, pressed=False, pressed_since=False, released_since=False
+    ),
+    enter=KeyState(
+        keypress=KP_ENTER, pressed=True, pressed_since=False, released_since=False
+    ),
+    exit=KeyState(
+        keypress=KP_EXIT, pressed=False, pressed_since=True, released_since=False
+    ),
+    left=KeyState(
+        keypress=KP_LEFT, pressed=False, pressed_since=False, released_since=True
+    ),
+    right=KeyState(
+        keypress=KP_RIGHT, pressed=True, pressed_since=True, released_since=False
+    ),
+    down=KeyState(
+        keypress=KP_DOWN, pressed=True, pressed_since=False, released_since=True
+    ),
 )